jakeLi0306

+++ repo/jcart-administration-back/src/main/java/io/ljh/jcartadministrationback/dto/in/OrderHistoryCreateInDTO.java
@@ -6,7 +6,7 @@ public class OrderHistoryCreateInDTO {
     private String comment;
     private Boolean customerNotified;
 
-    public Long getOrderId() {
+    public Integer getOrderId() {
         return orderId;
     }
